Fixed the Windows code that was messed up in an earlier change.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@184 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Engine/winstuff.cpp
@@ -142,6 +142,6 @@ void changeToUserDir () {
 }
 
 uint32_t launch(char * f) {
-	return ShellExecute (hMainWindow, "open", f, NULL, "C:\\", SW_SHOWNORMAL);
+	return (uint32_t) ShellExecute (hMainWindow, "open", f, NULL, "C:\\", SW_SHOWNORMAL);
 }
 #endif
